certificate-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
httpx==0.27.0
python-dateutil==2.9.0
asyncpg==0.29.0
prisma==0.15.0
python-dotenv==1.0.1
